Redesenhando núcleo de Delégua para trabalhar com outros dois núcleos: execução e tradução.
