+++ fontes/delegua.ts
@@ -1,615 +1,69 @@
-import * as sistemaArquivos from 'fs';
-import * as caminho from 'path';
-import * as readline from 'readline';
-import chalk from 'chalk';
-import colorizeJson from 'json-colorizer';
-
-import { Lexador } from '@designliquido/delegua/fontes/lexador/lexador';
-import { AvaliadorSintatico } from '@designliquido/delegua/fontes/avaliador-sintatico/avaliador-sintatico';
-import { Interpretador } from './interpretador/interpretador';
-
-import tiposDeSimbolos from '@designliquido/delegua/fontes/tipos-de-simbolos/delegua';
-
 import {
     AvaliadorSintaticoInterface,
-    InterpretadorComDepuracaoInterface,
-    InterpretadorInterface,
-    LexadorInterface,
-    RetornoExecucaoInterface,
-    SimboloInterface,
+    LexadorInterface
 } from '@designliquido/delegua/fontes/interfaces';
 
-import { InterpretadorEguaClassico } from '@designliquido/delegua/fontes/interpretador/dialetos/egua-classico/interpretador-egua-classico';
-import { LexadorEguaClassico } from '@designliquido/delegua/fontes/lexador/dialetos/lexador-egua-classico';
-import { LexadorPitugues } from '@designliquido/delegua/fontes/lexador/dialetos/lexador-pitugues';
-import { AvaliadorSintaticoPitugues } from '@designliquido/delegua/fontes/avaliador-sintatico/dialetos/avaliador-sintatico-pitugues';
-import { AvaliadorSintaticoEguaClassico, AvaliadorSintaticoPortugolIpt } from '@designliquido/delegua/fontes/avaliador-sintatico/dialetos';
-import { TradutorVisualg } from '@designliquido/delegua/fontes/tradutores/tradutor-visualg';
-
-import { Importador, RetornoImportador } from './importador';
-import { LexadorMapler } from '@designliquido/delegua/fontes/lexador/dialetos/lexador-mapler';
-import { AvaliadorSintaticoMapler } from '@designliquido/delegua/fontes/avaliador-sintatico/dialetos/avaliador-sintatico-mapler';
-import { LexadorVisuAlg } from '@designliquido/delegua/fontes/lexador/dialetos/lexador-visualg';
-import { AvaliadorSintaticoVisuAlg } from '@designliquido/delegua/fontes/avaliador-sintatico/dialetos/avaliador-sintatico-visualg';
-import { LexadorBirl } from '@designliquido/delegua/fontes/lexador/dialetos/lexador-birl';
-import { AvaliadorSintaticoBirl } from '@designliquido/delegua/fontes/avaliador-sintatico/dialetos/avaliador-sintatico-birl';
-import { TradutorJavaScript, TradutorPython, TradutorReversoJavaScript } from '@designliquido/delegua/fontes/tradutores';
-import { InterpretadorMapler } from '@designliquido/delegua/fontes/interpretador/dialetos/mapler/interpretador-mapler';
-import { InterpretadorVisuAlg } from '@designliquido/delegua/fontes/interpretador/dialetos/visualg/interpretador-visualg';
-import { ErroInterpretador } from '@designliquido/delegua/fontes/interpretador';
-import { InterpretadorBirl, InterpretadorPortugolIpt, InterpretadorPortugolStudio } from '@designliquido/delegua/fontes/interpretador/dialetos';
-import { InterpretadorPortugolStudioComDepuracao } from '@designliquido/delegua/fontes/interpretador/dialetos/portugol-studio/interpretador-portugol-studio-com-depuracao';
-import { LexadorPortugolStudio } from '@designliquido/delegua/fontes/lexador/dialetos/lexador-portugol-studio';
-import { AvaliadorSintaticoPortugolStudio } from '@designliquido/delegua/fontes/avaliador-sintatico/dialetos/avaliador-sintatico-portugol-studio';
-import { ServidorDepuracao } from './depuracao';
 import { DeleguaInterface, ImportadorInterface } from './interfaces';
-import { InterpretadorComDepuracaoImportacao } from './interpretador/interpretador-com-depuracao-importacao';
-import { InterpretadorVisuAlgComDepuracaoImportacao } from './interpretador/dialetos/interpretador-visualg-com-depuracao-importacao';
-import { InterpretadorMaplerComDepuracaoImportacao } from './interpretador/dialetos/interpretador-mapler-com-depuracao-importacao';
-import { LexadorPortugolIpt } from '@designliquido/delegua/fontes/lexador/dialetos';
-import { LexadorJson } from './lexador/lexador-json';
-import { FormatadorJson } from './formatadores';
+import { NucleoExecucao } from './nucleo-execucao';
+import { NucleoTraducao } from './nucleo-traducao';
 
 /**
  * O núcleo da linguagem.
  *
- * Responsável por avaliar a entrada fornecida, entender o código e executá-lo.
+ * Responsável por avaliar a entrada fornecida, chamar o núcleo 
+ * correspondente à operação solicitada e observar a execução.
  */
 export class Delegua implements DeleguaInterface {
-    dialeto: string;
-    dialetos: { [identificador: string]: string } = {
-        birl: 'BIRL',
-        delegua: 'padrão',
-        egua: 'Égua',
-        pitugues: 'ÉguaP',
-        mapler: 'mapler',
-        'portugol-studio': 'Portugol Studio',
-        visualg: 'VisuAlg',
-    };
-
-    arquivosAbertos: { [identificador: string]: string };
-    conteudoArquivosAbertos: { [identificador: string]: string[] };
-
-    interpretador: InterpretadorInterface;
-    lexador: LexadorInterface;
-    avaliadorSintatico: AvaliadorSintaticoInterface;
+    lexador: LexadorInterface<any>;
+    avaliadorSintatico: AvaliadorSintaticoInterface<any, any>;
     importador: ImportadorInterface;
-
-    traduzir: string = '';
-    tradutorJavaScript: TradutorJavaScript;
-    tradutorReversoJavascript: TradutorReversoJavaScript;
-    tradutorVisualg: TradutorVisualg;
-    tradutorPython: TradutorPython;
-    extensoes = {
-        delegua: '.delegua',
-        javascript: '.js',
-        js: '.js',
-        alg: '.alg',
-        visualg: '.alg',
-        python: '.py',
-        py: '.py',
-    }
 
     funcaoDeRetorno: Function;
     funcaoDeRetornoMesmaLinha: Function;
-    modoDepuracao: boolean;
-
-    servidorDepuracao: ServidorDepuracao;
 
     constructor(
-        dialeto = 'delegua',
-        performance = false,
-        depurador = false,
-        traduzir: any = null,
         funcaoDeRetorno: Function = null,
         funcaoDeRetornoMesmaLinha: Function = null
     ) {
-        this.arquivosAbertos = {};
-        this.conteudoArquivosAbertos = {};
-
-        this.dialeto = dialeto;
         this.funcaoDeRetorno = funcaoDeRetorno || console.log;
         // `process.stdout.write.bind(process.stdout)` é necessário por causa de 
         // https://stackoverflow.com/questions/28874665/node-js-cannot-read-property-defaultencoding-of-undefined
         this.funcaoDeRetornoMesmaLinha = funcaoDeRetornoMesmaLinha || process.stdout.write.bind(process.stdout);
-        this.modoDepuracao = depurador;
-
-        switch (this.dialeto) {
-            case 'birl':
-                this.lexador = new LexadorBirl();
-                this.avaliadorSintatico = new AvaliadorSintaticoBirl();
-                this.importador = new Importador(
-                    this.lexador,
-                    this.avaliadorSintatico,
-                    this.arquivosAbertos,
-                    this.conteudoArquivosAbertos,
-                    depurador
-                );
-                this.interpretador = new InterpretadorBirl(
-                    process.cwd(),
-                    this.funcaoDeRetorno,
-                    this.funcaoDeRetornoMesmaLinha);
-                break;
-            case 'egua':
-                if (depurador) {
-                    throw new Error('Dialeto ' + this.dialeto + ' não suporta depuração.');
-                }
-
-                this.lexador = new LexadorEguaClassico();
-                this.avaliadorSintatico = new AvaliadorSintaticoEguaClassico();
-                this.importador = new Importador(
-                    this.lexador,
-                    this.avaliadorSintatico,
-                    this.arquivosAbertos,
-                    this.conteudoArquivosAbertos,
-                    depurador
-                );
-                this.interpretador = new InterpretadorEguaClassico(process.cwd());
-                break;
-            case 'pitugues':
-                this.lexador = new LexadorPitugues();
-                this.avaliadorSintatico = new AvaliadorSintaticoPitugues();
-                this.importador = new Importador(
-                    this.lexador,
-                    this.avaliadorSintatico,
-                    this.arquivosAbertos,
-                    this.conteudoArquivosAbertos,
-                    depurador
-                );
-
-                this.interpretador = depurador
-                    ? new InterpretadorComDepuracaoImportacao(
-                        this.importador,
-                        process.cwd(),
-                        this.funcaoDeRetorno,
-                        this.funcaoDeRetornoMesmaLinha)
-                    : new Interpretador(this.importador,
-                        process.cwd(),
-                        performance,
-                        this.funcaoDeRetorno,
-                        this.funcaoDeRetornoMesmaLinha);
-                break;
-            case 'mapler':
-                this.lexador = new LexadorMapler();
-                this.avaliadorSintatico = new AvaliadorSintaticoMapler();
-                this.importador = new Importador(
-                    this.lexador,
-                    this.avaliadorSintatico,
-                    this.arquivosAbertos,
-                    this.conteudoArquivosAbertos,
-                    depurador
-                );
-
-                this.interpretador = depurador
-                    ? new InterpretadorMaplerComDepuracaoImportacao(
-                        this.importador,
-                        process.cwd(),
-                        this.funcaoDeRetorno,
-                        this.funcaoDeRetornoMesmaLinha)
-                    : new InterpretadorMapler(
-                        process.cwd(),
-                        false,
-                        this.funcaoDeRetorno,
-                        this.funcaoDeRetornoMesmaLinha);
-                break;
-            case 'portugol-ipt':
-                this.lexador = new LexadorPortugolIpt();
-                this.avaliadorSintatico = new AvaliadorSintaticoPortugolIpt();
-                this.importador = new Importador(
-                    this.lexador,
-                    this.avaliadorSintatico,
-                    this.arquivosAbertos,
-                    this.conteudoArquivosAbertos,
-                    depurador
-                );
-
-                this.interpretador = depurador
-                    ? new InterpretadorPortugolIpt(
-                        process.cwd(),
-                        this.funcaoDeRetornoMesmaLinha,
-                        this.funcaoDeRetornoMesmaLinha)
-                    : new InterpretadorPortugolIpt(
-                        process.cwd(),
-                        this.funcaoDeRetornoMesmaLinha,
-                        this.funcaoDeRetornoMesmaLinha);
-                break;
-            case 'portugol-studio':
-                this.lexador = new LexadorPortugolStudio();
-                this.avaliadorSintatico = new AvaliadorSintaticoPortugolStudio();
-                this.importador = new Importador(
-                    this.lexador,
-                    this.avaliadorSintatico,
-                    this.arquivosAbertos,
-                    this.conteudoArquivosAbertos,
-                    depurador
-                );
-
-                this.interpretador = depurador
-                    ? new InterpretadorPortugolStudioComDepuracao(
-                        process.cwd(),
-                        this.funcaoDeRetorno,
-                        this.funcaoDeRetornoMesmaLinha)
-                    : new InterpretadorPortugolStudio(
-                        process.cwd(),
-                        performance,
-                        this.funcaoDeRetorno);
-                break;
-            case 'visualg':
-                this.lexador = new LexadorVisuAlg();
-                this.avaliadorSintatico = new AvaliadorSintaticoVisuAlg();
-                this.importador = new Importador(
-                    this.lexador,
-                    this.avaliadorSintatico,
-                    this.arquivosAbertos,
-                    this.conteudoArquivosAbertos,
-                    depurador
-                );
-
-                this.interpretador = depurador
-                    ? new InterpretadorVisuAlgComDepuracaoImportacao(
-                        this.importador,
-                        process.cwd(),
-                        this.funcaoDeRetorno,
-                        this.funcaoDeRetornoMesmaLinha)
-                    : new InterpretadorVisuAlg(
-                        process.cwd(),
-                        false,
-                        this.funcaoDeRetorno,
-                        this.funcaoDeRetornoMesmaLinha);
-                break;
-            default:
-                this.lexador = new Lexador(performance);
-                this.avaliadorSintatico = new AvaliadorSintatico(performance);
-                this.importador = new Importador(
-                    this.lexador,
-                    this.avaliadorSintatico,
-                    this.arquivosAbertos,
-                    this.conteudoArquivosAbertos,
-                    depurador
-                );
-
-                this.interpretador = depurador
-                    ? new InterpretadorComDepuracaoImportacao(
-                        this.importador,
-                        process.cwd(),
-                        this.funcaoDeRetorno,
-                        this.funcaoDeRetornoMesmaLinha)
-                    : new Interpretador(
-                        this.importador,
-                        process.cwd(),
-                        performance,
-                        this.funcaoDeRetorno,
-                        this.funcaoDeRetornoMesmaLinha);
-                break;
-        }
-
-        if (traduzir) {
-            this.iniciarTradutor(traduzir)
-        }
-
-        if (depurador) {
-            this.iniciarDepuracao();
-        }
     }
 
-    iniciarTradutor(traduzir) {
-        this.traduzir = traduzir;
-        switch (traduzir) {
-            case 'delegua-para-js':
-            case 'delegua-para-javascript':
-                this.tradutorJavaScript = new TradutorJavaScript();
-                break;
-            case 'delegua-para-py':
-            case 'delegua-para-python':
-                this.tradutorPython = new TradutorPython();
-                break;
-            case 'js-para-delegua':
-            case 'javascript-para-delegua':
-                this.tradutorReversoJavascript = new TradutorReversoJavaScript();
-                break;
-            case 'alg-para-delegua':
-            case 'visualg-para-delegua':
-                this.tradutorVisualg = new TradutorVisualg();
-                break;
-            default:
-                throw new Error(`Tradutor '${traduzir}' não implementado.`);
-        }
+    async executarCodigoComoArgumento(
+        codigo: string,
+        dialeto: string = 'delegua',
+        performance: boolean = false
+    ): Promise<void> {
+        const nucleoExecucao = new NucleoExecucao(this.funcaoDeRetorno, this.funcaoDeRetornoMesmaLinha);
+        nucleoExecucao.configurarDialeto(dialeto, performance);
+        await nucleoExecucao.executarCodigoComoArgumento(codigo);
     }
 
-    versao(): string {
-        try {
-            const manifesto = caminho.resolve(__dirname, 'package.json');
-
-            return JSON.parse(sistemaArquivos.readFileSync(manifesto, { encoding: 'utf8' })).version || '0.23';
-        } catch (error: any) {
-            return '0.23 (desenvolvimento)';
-        }
+    async executarCodigoPorArquivo(
+        caminhoRelativoArquivo: string,
+        dialeto: string = 'delegua',
+        performance: boolean = false
+    ): Promise<any> {
+        const nucleoExecucao = new NucleoExecucao(this.funcaoDeRetorno, this.funcaoDeRetornoMesmaLinha);
+        nucleoExecucao.configurarDialeto(dialeto, performance);
+        await nucleoExecucao.carregarEExecutarArquivo(caminhoRelativoArquivo);
     }
 
-    /**
-     * LAIR (Leia-Avalie-Imprima-Repita) é o modo em que Delégua executa em modo console,
-     * ou seja, esperando como entrada linhas de código fornecidas pelo usuário.
-     */
-    iniciarLairDelegua(): void {
-        const lexadorJson = new LexadorJson();
-        const formatadorJson = new FormatadorJson();
-
-        this.funcaoDeRetorno(`Usando dialeto: ${this.dialetos[this.dialeto]}`);
-        this.funcaoDeRetorno(`Console da Linguagem Delégua v${this.versao()}`);
-        this.funcaoDeRetorno('Pressione Ctrl + C para sair');
-
-        const interfaceLeitura = readline.createInterface({
-            input: process.stdin,
-            output: process.stdout,
-            prompt: '\ndelegua> ',
-        });
-
-        const isto = this;
-
-        this.interpretador.interfaceEntradaSaida = interfaceLeitura;
-
-        interfaceLeitura.prompt();
-        interfaceLeitura.on('line', async (linha: string) => {
-            const { resultado } = await isto.executarUmaLinha(linha);
-            if (resultado && resultado.length) {
-                const resultadoLexacao = lexadorJson.getTokens(resultado[0]);
-                const resultadoFormatacao = formatadorJson.formatar(resultadoLexacao);
-                isto.funcaoDeRetorno(colorizeJson(resultadoFormatacao));
-            }
-
-            interfaceLeitura.prompt();
-        });
+    async iniciarLair(dialeto: string = 'delegua'): Promise<void> { 
+        const nucleoExecucao = new NucleoExecucao(this.funcaoDeRetorno, this.funcaoDeRetornoMesmaLinha);
+        nucleoExecucao.configurarDialeto(dialeto, false);
+        await nucleoExecucao.iniciarLairDelegua();
     }
 
-    /**
-     * Executa uma linha. Usado pelo modo LAIR e pelo servidor de depuração, quando recebe um comando 'avaliar'.
-     * @param linha A linha a ser avaliada.
-     * @returns O resultado da execução, com os retornos e respectivos erros, se houverem.
-     */
-    async executarUmaLinha(linha: string): Promise<RetornoExecucaoInterface> {
-        const retornoLexador = this.lexador.mapear([linha], -1);
-        const retornoAvaliadorSintatico = this.avaliadorSintatico.analisar(retornoLexador, -1);
-        if (
-            this.afericaoErros({
-                retornoLexador,
-                retornoAvaliadorSintatico,
-            } as RetornoImportador)
-        ) {
-            return { resultado: [] } as RetornoExecucaoInterface;
-        }
-
-        return await this.executar(
-            {
-                retornoLexador,
-                retornoAvaliadorSintatico,
-            } as RetornoImportador,
-            true
-        );
-    }
-
-    /**
-     * Instancia um servidor de depuração, normalmente recebendo requisições na porta 7777.
-     */
-    iniciarDepuracao(): void {
-        this.servidorDepuracao = new ServidorDepuracao(this);
-        this.servidorDepuracao.iniciarServidorDepuracao();
-        (this.interpretador as any).finalizacaoDaExecucao = this.finalizarDepuracao.bind(this);
-    }
-
-    /**
-     * Pede ao servidor de depuração que finalize a execução.
-     * Se não for feito, o servidor de depuração mantém um _stream_ aberto e nunca finaliza.
-     * Mais informações: https://stackoverflow.com/a/47456805/1314276
-     */
-    finalizarDepuracao(): void {
-        if (this.servidorDepuracao) {
-            this.servidorDepuracao.finalizarServidorDepuracao();
-        }
-    }
-
-    /**
-     * Verifica erros nas etapas de lexação e avaliação sintática.
-     * @param retornoImportador Um objeto que implementa a interface RetornoImportador.
-     * @returns Verdadeiro se há erros. Falso caso contrário.
-     */
-    afericaoErros(retornoImportador: RetornoImportador): boolean {
-        if (retornoImportador.retornoLexador.erros.length > 0) {
-            for (const erroLexador of retornoImportador.retornoLexador.erros) {
-                this.reportar(erroLexador.linha, ` no '${erroLexador.caractere}'`, erroLexador.mensagem);
-            }
-            return true;
-        }
-
-        if (retornoImportador.retornoAvaliadorSintatico.erros.length > 0) {
-            for (const erroAvaliadorSintatico of retornoImportador.retornoAvaliadorSintatico.erros) {
-                this.erro(erroAvaliadorSintatico.simbolo, erroAvaliadorSintatico.message);
-            }
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Realiza a tradução do arquivo passado como parâmetro no comando de execução.
-     * @param caminhoRelativoArquivo O caminho do arquivo.
-     * @param gerarArquivoSaida Se o resultado da tradução deve ser escrito em arquivo.
-     *                          Se verdadeiro, os arquivos de saída são escritos no mesmo diretório
-     *                          do arquivo passado no primeiro parâmetro.
-     */
-    traduzirArquivo(caminhoRelativoArquivo: string, gerarArquivoSaida: boolean): any {
-        const caminhoAbsolutoPrimeiroArquivo = caminho.resolve(caminhoRelativoArquivo);
-        const novoDiretorioBase = caminho.dirname(caminhoAbsolutoPrimeiroArquivo);
-
-        this.importador.diretorioBase = novoDiretorioBase;
-
-        const retornoImportador = this.importador.importar(
-            caminhoRelativoArquivo,
-            true,
-            !this.tradutorJavaScript && !this.tradutorPython
-        );
-
-        let resultado = null;
-        if (this.tradutorJavaScript !== null && this.tradutorJavaScript !== undefined) {
-            if (this.afericaoErros(retornoImportador)) {
-                process.exit(65); // Código para erro de avaliação antes da tradução
-            }
-
-            resultado = this.tradutorJavaScript.traduzir(retornoImportador.retornoAvaliadorSintatico.declaracoes);
-        } else if (this.tradutorPython !== null && this.tradutorPython !== undefined) {
-            if (this.afericaoErros(retornoImportador)) {
-                process.exit(65); // Código para erro de avaliação antes da tradução
-            }
-
-            resultado = this.tradutorPython.traduzir(retornoImportador.retornoAvaliadorSintatico.declaracoes);
-        } else if (this.tradutorReversoJavascript !== null && this.tradutorReversoJavascript !== undefined) {
-            resultado = this.tradutorReversoJavascript.traduzir(retornoImportador.conteudoArquivo.join('\n'));
-        } else {
-            resultado = this.tradutorVisualg.traduzir(retornoImportador.conteudoArquivo.join('\n'));
-        }
-
-        if (gerarArquivoSaida) {
-            const linguagem = this.traduzir?.split('-')[2] || '';
-            const extensaoAlvo = this.extensoes[linguagem]
-            if (extensaoAlvo) {
-                ['.delegua', '.js', '.alg'].map((extensao) => {
-                    if (caminhoAbsolutoPrimeiroArquivo.includes(extensao)) {
-                        sistemaArquivos.writeFile(caminhoAbsolutoPrimeiroArquivo.replace(extensao, `${extensaoAlvo}`), resultado, (erro) => {
-                            if (erro) throw erro;
-                        });
-                        return;
-                    }
-                });
-            }
-        }
-
-        this.funcaoDeRetorno(resultado);
-    }
-
-    async executarCodigoComoArgumento(codigo: string) {
-        const retornoLexador = this.lexador.mapear([codigo], -1);
-        const retornoAvaliadorSintatico = this.avaliadorSintatico.analisar(retornoLexador, -1);
-        const { erros } = await this.executar({
-            conteudoArquivo: [codigo],
-            nomeArquivo: '',
-            hashArquivo: -1,
-            retornoLexador: retornoLexador,
-            retornoAvaliadorSintatico: retornoAvaliadorSintatico,
-        });
-
-        if (erros.length > 0) process.exit(70); // Código com exceções não tratadas
-    }
-
-    /**
-     * Execução por arquivo.
-     * @param caminhoRelativoArquivo O caminho no sistema operacional do arquivo a ser aberto.
-     */
-    async carregarArquivo(caminhoRelativoArquivo: string): Promise<any> {
-        const caminhoAbsolutoPrimeiroArquivo = caminho.resolve(caminhoRelativoArquivo);
-        const novoDiretorioBase = caminho.dirname(caminhoAbsolutoPrimeiroArquivo);
-
-        this.importador.diretorioBase = novoDiretorioBase;
-        this.interpretador.diretorioBase = novoDiretorioBase;
-
-        const retornoImportador = this.importador.importar(caminhoRelativoArquivo, true, false);
-
-        if (this.afericaoErros(retornoImportador)) {
-            process.exit(65); // Código para erro de avaliação antes da execução
-        }
-
-        let errosExecucao: any = {
-            lexador: [],
-            avaliadorSintatico: [],
-            interpretador: [],
-        };
-
-        const interfaceLeitura = readline.createInterface({
-            input: process.stdin,
-            output: process.stdout,
-            prompt: '\n> ',
-        });
-
-        this.interpretador.interfaceEntradaSaida = interfaceLeitura;
-
-        if (this.modoDepuracao) {
-            try {
-                (this.interpretador as InterpretadorComDepuracaoInterface).prepararParaDepuracao(
-                    retornoImportador.retornoAvaliadorSintatico.declaracoes
-                );
-            } catch (erro: any) {
-                console.error(chalk.red(`[Erro de execução]`) + ` Dialeto ${this.dialeto} não suporta depuração.`);
-            }
-        } else {
-            const { erros } = await this.executar(retornoImportador);
-            errosExecucao = erros;
-        }
-
-        interfaceLeitura.close();
-        if (errosExecucao.length > 0) process.exit(70); // Código com exceções não tratadas
-    }
-
-    /**
-     * A execução do código de fato.
-     * @param retornoImportador Dados retornados do Importador, como o retorno do Lexador, do Avaliador
-     *                          Sintático e respectivos erros.
-     * @param manterAmbiente Indicação se ambiente deve ser mantido ou não. Normalmente verdadeiro
-     *                       para LAIR, falso para execução por arquivo.
-     * @returns Um objeto com o resultado da execução.
-     */
-    async executar(retornoImportador: RetornoImportador, manterAmbiente = false): Promise<RetornoExecucaoInterface> {
-        const retornoInterpretador = await this.interpretador.interpretar(
-            retornoImportador.retornoAvaliadorSintatico.declaracoes,
-            manterAmbiente
-        );
-
-        if (retornoInterpretador.erros.length > 0) {
-            for (const erroInterpretador of retornoInterpretador.erros) {
-                if (erroInterpretador.hasOwnProperty('simbolo')) {
-                    this.erroEmTempoDeExecucao(erroInterpretador);
-                } else {
-                    if (erroInterpretador.hasOwnProperty('erroInterno')) {
-                        const erroEmJavaScript: any = erroInterpretador as ErroInterpretador;
-                        console.error(
-                            chalk.red(`[Linha: ${erroEmJavaScript.linha}] Erro em JavaScript: `) +
-                            `${erroEmJavaScript.erroInterno?.message}`
-                        );
-                        console.error(chalk.red(`Pilha de execução: `) + `${erroEmJavaScript.erroInterno?.stack}`);
-                    } else {
-                        console.error(
-                            chalk.red(`Erro em JavaScript: `) + JSON.stringify(erroInterpretador)
-                        );
-                    }
-                }
-            }
-        }
-
-        return {
-            erros: retornoInterpretador.erros,
-            resultado: retornoInterpretador.resultado,
-        };
-    }
-
-    reportar(linha: number, onde: any, mensagem: string): void {
-        /* if (this.nomeArquivo)
-            console.error(
-                chalk.red(`[Arquivo: ${this.nomeArquivo}] [Linha: ${linha}]`) + ` Erro${onde}: ${mensagem}`
-            );
-        else */
-        console.error(chalk.red(`[Linha: ${linha}]`) + ` Erro${onde}: ${mensagem}`);
-    }
-
-    erro(simbolo: SimboloInterface, mensagemDeErro: string): void {
-        const _simbolo = simbolo || { tipo: tiposDeSimbolos.EOF, linha: -1, lexema: '(indefinido)' };
-        if (_simbolo.tipo === tiposDeSimbolos.EOF) {
-            this.reportar(Number(_simbolo.linha), ' no final do código', mensagemDeErro);
-        } else {
-            this.reportar(Number(_simbolo.linha), ` no '${_simbolo.lexema}'`, mensagemDeErro);
-        }
-    }
-
-    erroEmTempoDeExecucao(erro: any): void {
-        const linha = erro?.simbolo?.linha || erro?.linha;
-        const mensagem = erro?.mensagem || erro?.message;
-        console.error(chalk.red(`Erro: [Linha: ${linha}]`) + ` ${mensagem}`);
+    traduzirArquivo(
+        caminhoRelativoArquivo: string, 
+        comandoTraducao: string,
+        gerarArquivoSaida: boolean = false
+    ): void {
+        const nucleoTraducao = new NucleoTraducao(this.funcaoDeRetorno, this.funcaoDeRetornoMesmaLinha);
+        nucleoTraducao.iniciarTradutor(comandoTraducao);
+        nucleoTraducao.traduzirArquivo(caminhoRelativoArquivo, gerarArquivoSaida);
     }
 }
