+++ fontes/interfaces/delegua-interface.ts
@@ -1,30 +1,18 @@
-import { RetornoImportador } from '../importador';
-import {
-    AvaliadorSintaticoInterface,
-    InterpretadorComDepuracaoInterface,
-    InterpretadorInterface,
-    LexadorInterface,
-    RetornoExecucaoInterface,
-    SimboloInterface,
-} from '@designliquido/delegua/fontes/interfaces';
-import { ImportadorInterface } from './importador-interface';
-
 export interface DeleguaInterface {
-    dialeto: string;
-    arquivosAbertos: { [identificador: string]: string };
-    conteudoArquivosAbertos: { [identificador: string]: string[] };
-    funcaoDeRetorno: Function;
-
-    interpretador: InterpretadorInterface | InterpretadorComDepuracaoInterface;
-    lexador: LexadorInterface;
-    avaliadorSintatico: AvaliadorSintaticoInterface;
-    importador: ImportadorInterface;
-
-    versao(): string;
-    carregarArquivo(caminhoRelativoArquivo: string): void;
-    executar(retornoImportador: RetornoImportador): Promise<RetornoExecucaoInterface>;
-    executarUmaLinha(linha: string): Promise<RetornoExecucaoInterface>;
-    reportar(linha: number, onde: any, mensagem: string): void;
-    erro(simbolo: SimboloInterface, mensagemDeErro: string): void;
-    erroEmTempoDeExecucao(erro: any): void;
+    executarCodigoComoArgumento(
+        codigo: string,
+        dialeto: string,
+        performance: boolean
+    ): Promise<void>;
+    executarCodigoPorArquivo(
+        caminhoRelativoArquivo: string,
+        dialeto: string,
+        performance: boolean
+    ): Promise<any>;
+    iniciarLair(dialeto: string): Promise<void>;
+    traduzirArquivo(
+        caminhoRelativoArquivo: string, 
+        comandoTraducao: string,
+        gerarArquivoSaida: boolean
+    ): void;
 }
