+++ fontes/interfaces/nucleo-execucao-interface.ts
@@ -1,0 +1,11 @@
+import { InterpretadorInterface, RetornoExecucaoInterface } from "@designliquido/delegua/fontes/interfaces";
+
+export interface NucleoExecucaoInterface {
+    arquivosAbertos: { [identificador: string]: string };
+    conteudoArquivosAbertos: { [identificador: string]: string[] };
+
+    funcaoDeRetorno: Function;
+    interpretador: InterpretadorInterface;
+
+    executarUmaLinha(linha: string): Promise<RetornoExecucaoInterface>;
+}
